+++ src/main/java/br/com/giovanni/projetotg/service/GerenciaUsuarios.java
@@ -5,10 +5,13 @@ import br.com.giovanni.projetotg.dto.UsuarioDtoSummary;
 import br.com.giovanni.projetotg.model.Usuario;
 import br.com.giovanni.projetotg.repository.UsuarioRepository;
 import jakarta.persistence.EntityNotFoundException;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -41,7 +44,11 @@ public class GerenciaUsuarios {
     }
 
     public UsuarioDto editarUsuario(long id, String nome, String email, String password) {
+        String contextHolderEmail = Objects.requireNonNull(SecurityContextHolder.getContext().getAuthentication()).getName();
         Usuario usuario = usuarioRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Usuário não encontrado!"));
+        if (!usuario.getEmail().equalsIgnoreCase(contextHolderEmail)) {
+            throw new AccessDeniedException("Você não tem permissão para editar este usuário");
+        }
         if (nome != null && !nome.isBlank()) {
             usuario.setNome(nome);
         }
@@ -56,6 +63,11 @@ public class GerenciaUsuarios {
     }
 
     public void deletarUsuario(long id) {
+        String contextHolderEmail = Objects.requireNonNull(SecurityContextHolder.getContext().getAuthentication()).getName();
+        Usuario usuario = usuarioRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Usuário não encontrado!"));
+        if (!usuario.getEmail().equalsIgnoreCase(contextHolderEmail)) {
+            throw new AccessDeniedException("Você não tem permissão para deletar este usuário");
+        }
         usuarioRepository.deleteById(id);
     }
